Add auth guard

+++ my-app/src/app/components/crud/crud.component.ts
@@ -31,17 +31,19 @@ export class CrudComponent {
 
   addData() {
     if (this.formData?.id) {
-      this.userService
-        .updateUser(this.formData, this.formData.id)
-        .subscribe((res) => {
-          console.log('Update data successfully', res);
-          this.clearForm();
-          this.isSubmitedupd = true;
-          setTimeout(() => {
-            this.isSubmitedupd = false;
-          }, 2000);
-          this.fetchData();
-        });
+      if (confirm('Are you sure you want to update this item?')) {
+        this.userService
+          .updateUser(this.formData, this.formData.id)
+          .subscribe((res) => {
+            console.log('Update data successfully', res);
+            this.clearForm();
+            this.isSubmitedupd = true;
+            setTimeout(() => {
+              this.isSubmitedupd = false;
+            }, 2000);
+            this.fetchData();
+          });
+      }
     } else {
       this.userService.insertData(this.formData).subscribe((res) => {
         console.log('data Added Successfully');
@@ -56,22 +58,30 @@ export class CrudComponent {
   }
 
   deleteUser(id: string) {
-    this.userService.DeletetData(id).subscribe((res) => {
-      console.log('Data deleted successfully');
-      this.fetchData();
-    });
+    if (confirm('Are you sure you want to delete this item?')) {
+      this.userService.DeletetData(id).subscribe((res) => {
+        console.log('Data deleted successfully');
+        this.fetchData();
+      });
+    }
   }
-
 
   editUser(userData: any) {
     this.formData = { ...userData }; // Create a copy of the user object
   }
-  
-//this.formData = { ...userData }; // Creates a shallow copy
-//this.formData is a new object with the same properties but independent of userData.
-//Changes to form inputs do not affect the original userData in the table.
+
+  //this.formData = { ...userData }; // Creates a shallow copy
+  //this.formData is a new object with the same properties but independent of userData.
+  //Changes to form inputs do not affect the original userData in the table.
 
   clearForm(): void {
+    this.formData = {
+      name: '',
+      role: '',
+    };
+  }
+
+  cancel() {
     this.formData = {
       name: '',
       role: '',
